# core/openapi.yaml
openapi: 3.0.3
info:
  title: Pen CMS API Specification
  description: |
    The contract-first API specification for Pen CMS, a Git-native flat-file CMS.

    Humans and agents share one capability vocabulary (string lists). `write:*` is
    create+update (never split create:/update:). `delete:*` is separate. `read` is
    monolithic. `publish:content` is content status (PATCH approve/publish) and is
    not host `publish` (deploy + Deploy Grant). Generic PUT of a post/page does not
    sniff status — that stays under write:posts / write:pages.

    Legacy agent scope `write` expands one-way to every write:*, every delete:*, and
    publish:content. It never implies host publish, users:manage, or manage:sites.

    Human admin routes under /api authorize from cookie `pen_jwt` or Bearer JWT →
    user YAML. Forgeable `pen_role` / `pen_user_id` cookies are not a root of trust.
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /publish/providers:
    servers:
      - url: /api
    get:
      summary: List publish provider catalog
      description: >
        Core catalog is SFTP and GitHub Pages. Pro overlay registers additional
        cloud adapters. Each entry includes ui_schema for the admin Publish tab.
      security:
        - CookieJwtAuth: []
        - BearerJwt: []
      responses:
        '200':
          description: Provider catalog.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishProviderCatalog'
  /publish/target:
    servers:
      - url: /api
    get:
      summary: Read publish target
      description: Non-secret host config for a site. Secrets stay in the ZK vault.
      security:
        - CookieJwtAuth: []
        - BearerJwt: []
      parameters:
        - name: site
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Publish target (configured or unconnected).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTarget'
    put:
      summary: Save publish target
      description: >
        Persist non-secret host settings. Passwords and tokens must not be sent.
        Core providers are sftp and github_pages. Unregistered provider ids return 400.
      security:
        - CookieJwtAuth: []
        - BearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishTargetPut'
      responses:
        '200':
          description: Saved target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTarget'
        '400':
          description: Invalid payload, secret keys, or unknown provider.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Pen-API-Key
    CookieJwtAuth:
      type: apiKey
      in: cookie
      name: pen_jwt
      description: Human session JWT. Authorize from this cookie → YAML. Never trust pen_role / pen_user_id.
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human or agent JWT. Agent tokens use type=agent + scopes + site_id; they do not inherit the sponsor admin role.
...
  schemas:
...
    PublishProviderCatalog:
      type: object
      required:
        - providers
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/PublishProvider'
    PublishProvider:
      type: object
      required:
        - id
        - label
        - enabled
      properties:
        id:
          type: string
          description: Core ids are sftp and github_pages.
        label:
          type: string
        enabled:
          type: boolean
        capabilities:
          type: object
        yaml_fields:
          type: array
          items:
            type: string
        vault_key:
          type: string
        http_alias:
          type: string
        ui_schema:
          type: object
    PublishTarget:
      type: object
      properties:
        site_id:
          type: string
        configured:
          type: boolean
        provider:
          type: string
          nullable: true
        host:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        username:
          type: string
          nullable: true
        remote_path:
          type: string
          nullable: true
        auth_method:
          type: string
          nullable: true
        public_url:
          type: string
          nullable: true
        last_published_at:
          type: string
          nullable: true
        last_status:
          type: string
          nullable: true
        agent_publish:
          type: string
          nullable: true
        github_owner:
          type: string
          nullable: true
        github_repo:
          type: string
          nullable: true
        github_pages_branch:
          type: string
          nullable: true
        github_pages_cname:
          type: string
          nullable: true
        webhook_url:
          type: string
          nullable: true
        has_webhook_secret:
          type: boolean
    PublishTargetPut:
      type: object
      required:
        - site
      properties:
        site:
          type: string
        provider:
          type: string
          description: Core values are sftp and github_pages.
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        remote_path:
          type: string
        auth_method:
          type: string
        public_url:
          type: string
        github_owner:
          type: string
        github_repo:
          type: string
        github_pages_branch:
          type: string
        github_pages_cname:
          type: string
        webhook_url:
          type: string
        webhook_secret:
          type: string
          description: Write-only HMAC signing secret; never returned on GET.